+++ components/telemetry-operator/internal/controller/logpipeline/sync/mocks/Client.go
