+++ components/telemetry-operator/internal/controller/logpipeline/sync/sync.go
